% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBalance.R
\name{checkBalance}
\alias{checkBalance}
\title{checkBalance}
\usage{
checkBalance(data, ..., design, x, size = 3)
}
\arguments{
\item{data}{A data frame output from any easyXpress function after the \code{summarizeWells} function.}

\item{...}{<[rlang::`dyn-dots`]> Variable(s) used to facet the data in the output plot. Variable names can be listed in succession.}

\item{design}{A design data frame like the one output from the \code{readXpress} function or one read into R from the raw design file(s).
All the variables specified in \code{...} and \code{x} must be present in the design data frame.}

\item{x}{The independent variable to use as the x-axis of the plot. For example, strain. No quotes are needed.}

\item{size}{The size of the number label used for plotting. The default is 3.}
}
\value{
A list containing two elements.
1) A data frame used to make the diagnostic plot
2) A plot showing the data completeness post filtering. It is faceted by \code{...}.
}
\description{
A function to visualize experimenta balance after filtering objects and wells.
This function requires a design data frame.
}
