% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEff.R
\name{checkEff}
\alias{checkEff}
\title{checkEff}
\usage{
checkEff(data, ..., x, y, fill = NULL, size = 1.5, scales = "fixed")
}
\arguments{
\item{data}{A data frame output from any easyXpress function used after the \code{modelSelection} function.}

\item{...}{<[rlang::`dyn-dots`]> Variable(s) used to facet the data in the output plot. Variable names can be listed in succession.}

\item{x}{The independent variable to use as the x-axis of the plot. For example, strain. No quotes are needed.}

\item{y}{The dependent variable to use as the y-zxis of the plot. For example, median_wormlength_um. No quotes are needed.}

\item{fill}{The variable used to fill data points.}

\item{size}{The size of the points plotted with \code{ggplot2::geom_jitter}. The default is 1.5.}

\item{scales}{How to plot axes for facets. Should scales be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")?}
}
\value{
A box plot of the data with points jittered behind. The plot is faceted by any variables supplied with ...,
}
\description{
A function to visualize experimental effects and detect potential issues in the data.
}
