% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkModelsView.R
\name{checkModelsView}
\alias{checkModelsView}
\title{checkModelsView}
\usage{
checkModelsView(
  data,
  ...,
  modelName = "MDHD",
  OF = "filter",
  length_thresh = 164.685,
  strainN = 4,
  wellN = 4,
  proc.img.dir,
  well.label,
  out.dir
)
}
\arguments{
\item{data}{A data frame output from the \code{modelSelection} function or any \code{OF} function from easyXpress.}

\item{...}{Variable(s) used to group data. Variable names in data are supplied separated by commas and without quotes. For example \code{drug, concentration_um}.}

\item{modelName}{The name of the cellprofiler model to check in the diagnostic plots. Typically this is the smallest model in the dataset.
The default is set to \code{"MDHD"}.}

\item{OF}{Select one of \code{"filter", "ignore"}. The default is \code{"filter"}, which will filter out all objects flagged by OFs.
\code{"ignore"} will include all objects in the diagnostic plots.}

\item{length_thresh}{An object length threshold in um used to label objects from the data. The default setting is \code{164.685} um.
This is the standard threshold used for the AndersenLab 2X objective images. Please adjust if necessary.}

\item{strainN}{The maximum number of strains to include in a overlay. This function will return at maximum \code{strainN} strains per overlay.
These are strains are choosen because they have the lowest average object length in the group.}

\item{wellN}{The maximum number of wells to include in a overlay. This function will return at maximum \code{wellN} wells per strain within an overlay.
These wells are choosen because they have the lowest average object length within the strain.}

\item{proc.img.dir}{Supply one of two options. 1 - The full path to a directory holding all processed images. For example, "~/proc_images/".
2 - A variable name in \code{data} that holds the full path to the directory holding processed images matching the data.
Users will need to add this variable to the data frame themselves. For example, "image_path".
For either option to work the processed images must have the standard \code{_overlay.png} suffix and file name output from CellProfiler.}

\item{well.label}{A variable name in \code{data} to display as a well label. For example, "Metadata_Well".}

\item{out.dir}{The full path to an existing directory that will be used to hold the overlays.}
}
\value{
A dataframe nested by the grouping variables with or without flagged objects removed.
}
\description{
This function will create an one or more arrays of image overlays to help users evaluate cellprofiler model performance and object length thresholds.
}
