% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkObjs.R
\name{checkObjs}
\alias{checkObjs}
\title{checkObjs}
\usage{
checkObjs(data, OF, ...)
}
\arguments{
\item{data}{A data frame output from the \code{modelSelection} function or any \code{OF} function from easyXpress.}

\item{OF}{Select one of \code{"filter", "label", "ignore"}. The default is \code{"filter"}, which will filter out all objects flagged by OFs.
\code{"label"} will fill flagged objects by OFs and fill objects not flagged with model names. \code{"ignore"} will fill all objects by model names.}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> Variable(s) used to summarize data. Variable names can be listed in succession.}
}
\value{
A plot faceted by the grouping variables ..., with objects labelled by the CellProfiler worm model and OFs if specified.
}
\description{
A function to help detect potential issues in the data by visualizing the size and type of objects across grouping variables.
}
