% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkWF.R
\name{checkWF}
\alias{checkWF}
\title{checkWF}
\usage{
checkWF(data, ..., plot = T)
}
\arguments{
\item{data}{A data frame output from any WellFlag (WF) function from easyXpress.}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> Variable(s) used to group data. Variable names in data are supplied separated by commas and without quotes.}

\item{plot}{Logical, if \code{TRUE}, the defualt, a ggplot2 well will be returned in a list.
For example, \code{drug, strain}.}
}
\value{
If \code{plot = TRUE} a list with two elements is returned. The first element (d) is a data frame summarized by ...,
with the numbers of wells flagged and retained by each WF applied to the data. The second element (p)
is a ggplot2 object showing the wells retained after all filtering and faceted by ...
If \code{plot = FALSE} only a data frame is returned.
}
\description{
checkWF
}
