% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterOF.R
\name{clusterOF}
\alias{clusterOF}
\title{clusterOF}
\usage{
clusterOF(data)
}
\arguments{
\item{data}{A data frame output from the \code{modelSelection} and \code{edgeOF} functions.}
}
\value{
A single data frame identical to the input data with the \code{cluster_ObjectFlag} variable added.
The \code{cluster_ObjectFlag} variable is coded as \code{"cluster"} for objects that are clustered,
all other objects are coded as \code{NA_character}.
}
\description{
This function will flag objects that belong to the same \code{Parent_WormObject} identified by cellprofiler.
}
