% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyXpress-package.R
\docType{package}
\name{easyXpress-package}
\alias{easyXpress}
\alias{easyXpress-package}
\title{easyXpress}
\description{
A workflow for the reading, processing, and visualization
    of data obtained from the Molecular Devices ImageExpress Nano Imager.
    Includes a powerful suite of functions for the rapid processing
    and analysis of large high-throughput image based datasets.
}
\details{
add package details  here
}
\author{
\strong{Maintainer}: Joy Nyaanga \email{jnyaan@gmail.com} (\href{https://orcid.org/0000-0002-1402-9213}{ORCID})

Authors:
\itemize{
  \item Tim Crombie \email{tim.crombie@gmail.com} (\href{https://orcid.org/0000-0002-5645-4154}{ORCID})
  \item Sam Widmayer \email{sjwidmay@gmail.com} (\href{https://orcid.org/0000-0002-1200-4768}{ORCID})
}

}
\keyword{internal}
