% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeOF.R
\name{edgeOF}
\alias{edgeOF}
\title{edgeOF}
\usage{
edgeOF(data, radius = 825, center_x = 1024, center_y = 1024)
}
\arguments{
\item{data}{A data frame output from the \code{modelSelection} function.}

\item{radius}{Radius in pixels away from image center with even illumination. The default
value is set to 825, whcih is standard for images taken with the imageXpress nano 2X objective.}

\item{center_x}{center x position of image. Default is 1024.}

\item{center_y}{center y position of image. Default is 1024.}
}
\value{
A single data frame with \code{edge_ObjectFlag} and \code{well_edge_flag_radius}
variables added. The \code{edge_ObjectFlag} variable is coded as \code{"edge"} for objects with centroids outside
the radius provided, all other objects are coded as \code{NA_character}.
}
\description{
This function will flag objects on the edge of wells that are difficult to segement with
CellProfiler.
}
