% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOF.R
\name{filterOF}
\alias{filterOF}
\title{filterOF}
\usage{
filterOF(data, rmVars = TRUE)
}
\arguments{
\item{data}{A data frame with \code{_ObjectFlag} variables. Typically the data is generated with easyXpress OF functions.}

\item{rmVars}{Logical, if \code{TRUE} the \code{_ObjectFlag} variables are removed from the data frame after flagged objects are removed.
If \code{FALSE}, the variables are retained. This argument can be useful for joining data frames. The default is \code{TRUE}.}
}
\value{
A single data frame with all flagged objects removed.
}
\description{
This function will filter out all flagged objects identified by \code{_ObjectFlag} variables.
}
