% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterWF.R
\name{filterWF}
\alias{filterWF}
\title{filterWF}
\usage{
filterWF(data, rmVars = TRUE)
}
\arguments{
\item{data}{A data frame with \code{_WellFlag} variables. Typically the data is generated with easyXpress WF functions.}

\item{rmVars}{Logical, if \code{TRUE} the \code{_WellFlag} variables are removed from the data frame after flagged objects are removed.
If \code{FALSE}, the variables are retained. This argument can be useful for joining data frames. The default is \code{TRUE}.}
}
\value{
A single data frame with all flagged wells removed.
}
\description{
This function will filter out all flagged wells identified by \code{_WellFlag} variables.
}
