% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqrOutlier.R
\name{iqrOutlier}
\alias{iqrOutlier}
\title{iqrOutlier}
\usage{
iqrOutlier(x, na.rm = TRUE, thresh)
}
\arguments{
\item{x}{A numerical vector to perform outlier removal on.}

\item{na.rm}{Logical, if \code{TRUE}, NAs are removed from the claculation.}

\item{thresh}{Numeric, a constant to multiply the IQR by for detecting outiers.}
}
\value{
A vector with outliers set the NA.
}
\description{
This function will identify outliers using the interquartile range (IQR).
}
