% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDesign.R
\name{makeDesign}
\alias{makeDesign}
\title{makeDesign}
\usage{
makeDesign(
  proj,
  n.plate,
  n.row = 8,
  n.col = 12,
  assay.type = NULL,
  food = NULL,
  od = NULL
)
}
\arguments{
\item{proj}{The name of the project this design file will be used for. Remember that the project name always
follows the 8-digit date in project directories. For example, the project directory \code{20200130_toxin01A}
has a project name of \code{"toxin01A"}. The project name will be used to join the design file to the
\code{cellprofiler-nf} output data by the \code{Metadata_Experiment} variable, so it is critical that they
match exactly.}

\item{n.plate}{The number of plates in the experiment.}

\item{n.row}{The number of rows in the plates. Default is 8 for a 96-well plate.}

\item{n.col}{The number of columns in the plates. Default is 12 for a 96-well plate.}

\item{assay.type}{OPTIONAL: The type of assay run. For example, "48h".}

\item{food}{OPTIONAL: A character description of the food used to perform the assay. For example, "15hHB101_20220727".}

\item{od}{OPTIONAL: The final optical density of bacterial food used in the assay. For example, 10.}
}
\value{
A single data frame with the variables \code{"Metadata_Plate"} and \code{"Metadata_Well"} variables that can be used
as a design file for the \code{readXpress} function. The data frame can be augmented further to include specific drugs and dilutents
or other experimental design attributes if desired. The variables output include:
}
\description{
This function will aid in making a design file for an imaging experiment (design.csv).
}
\details{
The variables output are as follows:\tabular{lc}{
   Variables \tab Class \cr
   Metadata_Experiment \tab chr \cr
   Metadata_Plate \tab chr \cr
   Metadata_Well \tab chr \cr
   strain \tab chr \cr
   drug \tab chr \cr
   concentration_um \tab dbl \cr
   bleach \tab chr \cr
   diluent \tab chr \cr
   assay_type \tab chr \cr
   plate \tab dbl \cr
   row \tab chr \cr
   col \tab chr \cr
   food \tab chr \cr
   od \tab int \cr
   well_censor \tab chr \cr
   well_censor_reason \tab chr \cr
   notes \tab chr \cr
}
}
