% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{model_select_1}
\alias{model_select_1}
\title{Generalized model for one worm object}
\format{
A dataframe with 3 rows and 4 variables:
\describe{
  \item{m1}{number of objects identified as model number 1}
  \item{model_select}{final selected model}
  \item{model_flag}{flag classifier}
  \item{cluster_flag}{indicates presence of a cluster}
}
}
\description{
This data frame conains the generalized model selection dataframe
for a one worm model CellProfiler run.
}
