% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{model_select_2}
\alias{model_select_2}
\title{Generalized model for two worm objects}
\format{
A dataframe with 9 rows and 5 variables:
\describe{
  \item{m1}{number of objects identified as model number 1}
  \item{m2}{number of objects identified as model number 2}
  \item{model_select}{final selected model}
  \item{model_flag}{flag classifier}
  \item{cluster_flag}{indicates presence of a cluster}
}
}
\description{
This data frame conains the generalized model selection dataframe
for a two worm model CellProfiler run.
}
