% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{model_select_3}
\alias{model_select_3}
\title{Generalized model for three worm objects}
\format{
A dataframe with 27 rows and 6 variables:
\describe{
  \item{m1}{number of objects identified as model number 1}
  \item{m2}{number of objects identified as model number 2}
  \item{m3}{number of objects identified as model number 3}
  \item{model_select}{final selected model}
  \item{model_flag}{flag classifier}
  \item{cluster_flag}{indicates presence of a cluster}
}
}
\description{
This data frame conains the generalized model selection dataframe
for a three worm model CellProfiler run.
}
