% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{model_select_4}
\alias{model_select_4}
\title{Generalized model for four worm objects}
\format{
A dataframe with 81 rows and 7 variables:
\describe{
  \item{m1}{number of objects identified as model number 1}
  \item{m2}{number of objects identified as model number 2}
  \item{m3}{number of objects identified as model number 3}
  \item{m4}{number of objects identified as model number 4}
  \item{model_select}{final selected model}
  \item{model_flag}{flag classifier}
  \item{cluster_flag}{indicates presence of a cluster}
}
}
\description{
This data frame contains the generalized model selection dataframe
for a four worm model CellProfiler run.
}
