% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{modelSelection}
\alias{modelSelection}
\title{modelSelection}
\usage{
modelSelection(df)
}
\arguments{
\item{df}{A single dataframe containing worm object measurements
from each CellProfiler model used in an asssay. A model variable is required.}
}
\value{
A single data frame named model_selected that contains
the best CellProfiler model for detecting worm objects within
each primary object detected by CellProfiler.
}
\description{
This function will assign the appropriate CellProfiler model
to each primary object in a raw_data object output
by the readXpress function.
}
