% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nWF.R
\name{nWF}
\alias{nWF}
\title{nWF}
\usage{
nWF(data, ..., max = 30, min = 5, plot = T)
}
\arguments{
\item{data}{A data frame output from the \code{summarizeWells} function.}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> Variable(s) used to group data for ploting. Variable names in data are supplied separated by commas and without quotes.}

\item{max}{A numeric value used to flag wells with too many objects. The default is \code{30}.}

\item{min}{A numeric value used to flag wells with too few objects. The default is \code{5}.}

\item{plot}{Logical, if \code{TRUE}, the default. A ggplot2 object will be returned showing the distribution of object counts in wells (n).}
}
\value{
Either a single data frame identical to the input data with the \code{n_WellFlag} variable added.
The \code{n_WellFlag} variable is coded as \code{"n<min"} or \code{n>max} for wells with object counts (n) outside the range set by \code{max} and \code{min}.
Or, if \code{plot = T}, a list with two elements, the first element is the data frame, the other is the diagnostic plot.
}
\description{
This function will flag wells that have too many or too few objects.
}
