% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierOF.R
\name{outlierOF}
\alias{outlierOF}
\title{outlierOF}
\usage{
outlierOF(data, iqr = TRUE, thresh = NULL, filterOF = TRUE)
}
\arguments{
\item{data}{A data frame output from the \code{modelSelection} function.}

\item{iqr}{Logical, if \code{TRUE}, objects in a well are called outliers if their worm_length_um is outside the range \code{median(worm_length_um) +/- (thresh * IQR)}.
If \code{FALSE}, objects in a well are called outliers if their worm_length_um is outside the range \code{mean(worm_length_um) +/- (thresh * SD)}.
The default is \code{TRUE}. NOTE: The Interquartile Range (IQR) method is useful for most distributions. The Standard Deviation (SD) method should only
be applied to normal distributions.}

\item{thresh}{A numeric value used to calculate outliers. By default this parameter is set to \code{NULL}.
If \code{thresh = NULL} then it is set to the recommended values within the function:
(\code{1.5} if \code{iqr = TRUE}) and (\code{3} if \code{iqr = FALSE}).}

\item{filterOF}{Logical, if \code{TRUE} any flagged objects detected in data are not considered for the calculation of outliers.}
}
\value{
A single data frame identical to the input data with the \code{outlier_ObjectFlag} variable added.
The \code{outlier_ObjectFlag} variable is coded as \code{"outlier"} for objects deemed outliers and \code{NA_character_} for non-outliers.
}
\description{
This function will flag outlier objects based the Worm_Legnth variable.
}
