% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierWF.R
\name{outlierWF}
\alias{outlierWF}
\title{outlierWF}
\usage{
outlierWF(data, ..., iqr = TRUE, thresh = NULL, filterWF = TRUE)
}
\arguments{
\item{data}{A data frame output from the \code{summarizeWells}function or any of the \code{WF} functions.}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> Variable(s) used to group the well summarized data. Variable names in data are supplied separated by commas and without quotes.
The outliers will be detected across the wells within the group specified.}

\item{iqr}{Logical, if \code{TRUE}, wells in a group are called outliers if the well median_wormlength_um is outside the range \code{median(median_wormlength_um) +/- (thresh * IQR)}.
If \code{FALSE}, wells in a group are called outliers if the well median_wormlength_um is outside the range \code{mean(median_wormlength_um) +/- (thresh * SD)}.
The default is \code{TRUE}. NOTE: The Interquartile Range (IQR) method is useful for most distributions. The Standard Deviation (SD) method should only
be applied to normal distributions.}

\item{thresh}{A numeric value used to calculate outliers. By default this parameter is set to \code{NULL}.
If \code{thresh = NULL} then it is set to the recommended values within the function:
(\code{1.5} if \code{iqr = TRUE}) and (\code{3} if \code{iqr = FALSE}).}

\item{filterWF}{Logical, if \code{TRUE} any flagged wells detected in data are not considered for the calculation of outliers.}
}
\value{
A single data frame identical to the input data with the \code{outlier_WellFlag} variable added.
The \code{outlier_WellFlag} variable is coded as \code{"outlier"} for objects deemed outliers and \code{NA_character_} for non-outliers.
}
\description{
This function will flag outlier wells based the median_wormlength_um for wells.
}
