% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateThumbs.R
\name{plateThumbs}
\alias{plateThumbs}
\title{plateThumbs}
\usage{
plateThumbs(project_dir, plates = "all", file_ext = ".TIF")
}
\arguments{
\item{project_dir}{The full path to the project directory. It must contain
the raw_image_thumbs folder holding .png thumbnail images of wells. The
\code{wellThumbs} function can be used to generate the raw_image_thumbs
folder and .png thumbnails from raw .TIFs.}

\item{plates}{a vector with plate numbers to make thumbs for. This vector is
used to match plate patterns in the filelist.\cr \code{"all"} will make
thumbs for all plates.\cr \code{13:72} will make thumbs for wells plates
p013 - p072.\cr \code{c(1, 3, 106)} will make thumbs for plates p001, p003, and
p106.}

\item{file_ext}{the file extension of your raw images. The default is \code{".TIF"},
this argument is case sensitive. This means .tif files will not be found by default.}
}
\value{
A folder named raw_plate_thumbs under the project directory
  containing individual .png files for each plate.
}
\description{
Make a full 96-well plate image in .png format from well thumbnail images.
}
