% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{process}
\usage{
process(flag_data, ...)
}
\arguments{
\item{flag_data}{Flagged data to be summarized}

\item{...}{specify variable used to summarize data}
}
\value{
List with four elements: raw data, processed data, and summaries for both datasets.
}
\description{
process
}
