% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regEff.R
\name{regEff}
\alias{regEff}
\title{regEff}
\usage{
regEff(data, ..., d.var, c.var)
}
\arguments{
\item{data}{A data frame output from any easyXpress function used after the \code{modelSelection} function.}

\item{...}{<[rlang::`dyn-dots`]> Variable(s) used to group the data prior to regression. Variable names can be listed in succession. For example, \code{drug}.}

\item{d.var}{The dependent variable to use in the linear model. Supply the variable without quotes. For example, \code{median_wormlength_um_delta}.}

\item{c.var}{The confounding variable to use in the linear model. Supply the variable without quotes. For example, \code{bleach}.}
}
\value{
A list including four elements.
1) A data frame, \code{<out>$d}, with three variables added. These include, \code{<d.var>_reg},
\code{<c.var>_reg_coeff}, and \code{<c.var>_reg_sig}.
2) A diagnostic plot, \code{<out>$p1}, showing the effect of the confounding variable (\code{c.var}) on the dependent variable (\code{d.var}).
3) A diagnositc plot, \code{<out>$p2}, showing the regression coefficients of the confounding variable (\code{c.var}) on the y-axis and the dependent variable (\code{d.var}) on the x-axis.
4) A list, \code{<out>$models} of model objects for each group.
}
\description{
A function to regress out the effect a confounding variable on a dependent variable.
}
