% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdOutlier.R
\name{sdOutlier}
\alias{sdOutlier}
\title{sdOutlier}
\usage{
sdOutlier(x, na.rm = TRUE, thresh)
}
\arguments{
\item{x}{A numerical vector to perform outlier removal on.}

\item{na.rm}{Logical, if \code{TRUE}, NAs are removed from the claculation.}

\item{thresh}{Numeric, a constant to multiply the SD by for detecting outiers.}
}
\value{
A vector with outliers set the NA.
}
\description{
This function will identify outliers using the standard deviation.
}
