% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFlags.R
\name{setFlags}
\alias{setFlags}
\title{setFlags}
\usage{
setFlags(data, cluster_flag = TRUE, well_edge_flag = TRUE)
}
\arguments{
\item{data}{CellProfiler data following model selection and edge flagging}

\item{cluster_flag}{Logical parameter; do you want worm objects
in a cluster to be excluded when calculating well outliers?
We recommend TRUE as the default.}

\item{well_edge_flag}{Logical parameter; do you want worm objects
in close proximity to the well edge to be excluded
when calculating well outliers? We recommend TRUE as the default.}
}
\value{
A single data frame named raw_data that contains
all CellProfiler model outputs and experimental treatments
if a design file is used.
}
\description{
setFlags
}
