% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOF.R
\name{setOF}
\alias{setOF}
\title{setOF}
\usage{
setOF(data)
}
\arguments{
\item{data}{A data frame with \code{_ObjectFlag} variables. Typically the data is generated with easyXpress OF functions but any variable
that contains \code{_ObjectFlag} in its name will be considered.}
}
\value{
A single data frame with all flagged objects labelled in the \code{objectFlag} variable. All objects that are not flagged are
labelled \code{noFlag}.
}
\description{
This function will assign all flagged objects a single value returned as the \code{objectFlag} variable.
It is useful for collapsing many easyXpress OFs into a single flag label based on the order in which the flags were
added to the data frame.
}
