% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWF.R
\name{setWF}
\alias{setWF}
\title{setWF}
\usage{
setWF(data)
}
\arguments{
\item{data}{A data frame with \code{_WellFlag} variables. Typically the data is generated with easyXpress WF functions but any variable
that contains \code{_WellFlag} in its name will be considered.}
}
\value{
A single data frame with all flagged wells labelled in the \code{wellFlag} variable. All objects that are not flagged are
labelled \code{noFlag}.
}
\description{
This function will assign all flagged wells a single value returned as the \code{wellFlag} variable.
It is useful for collapsing many easyXpress WFs into a single flag label based on the order in which the flags were
added to the data frame.
}
