% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeWells.R
\name{summarizeWells}
\alias{summarizeWells}
\title{summarizeWells}
\usage{
summarizeWells(data, OF = "filter", drop = T)
}
\arguments{
\item{data}{A data frame output from the \code{model_selection} or \code{OF} functions.}

\item{OF}{Specify either \code{"filter"} or \code{"ignore"}. The default is \code{"filter"}, which will filter out flagged objects.
\code{"ignore"} will ignore flagged data and include those objects in the summary statistics for a well. Setting OF to \code{"ignore"} is NOT recommended.}

\item{drop}{Logical, if \code{TRUE}, the default. The standard variables holding object data are dropped.
Be careful using \code{FALSE}. If you plan to return to using the object level data, only the variables will be retained but all rows (objects) will not be present.
Therefore, the unsummarized object data will need to be rejoined with the summarized data to have both present in a single dataframe. The best practice is to retain the object level data in one data frame
and summarized well data in another, which can be used with the well flag (WF) functions to complete the processing.}
}
\value{
A data frame with summary statistics for wells to be used with easyXpress well flag (WF) functions.
}
\description{
Summarize the objects within wells by calculating various population statistics, including the median, mean, standard deviation, and coefficient of variation (CV) for \code{worm_length_um}.
The number of objects in the wells are also calculated. The output can be used with the various well flag (WF) functions if desired.
}
