% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewDose.R
\name{viewDose}
\alias{viewDose}
\title{viewDose}
\usage{
viewDose(data, strain_name, drug_name, proc_img_dir)
}
\arguments{
\item{data}{Unsummarised CellProfiler data output from the
\code{Xpress} functon. Either the raw or processed data can be viewed.}

\item{strain_name}{A character value for the strain to use in
the plots, e.g. \code{"N2"}.}

\item{drug_name}{A character value for the drug to use in
the plots, e.g. \code{"paraquat"}.}

\item{proc_img_dir}{The full path of the directory holding
the processed images.}
}
\value{
A ggplot object with representative wells for each dose
of the selected drug and strain. The user cannot specify the
individual wells to be used. To view the results for a specific
well use the \code{viewWell} function.
}
\description{
#'\code{viewDose} plots representative processed well images
with object centroids colored by type for each concentration
of a drug.
}
