% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewOverlay.R
\name{viewOverlay}
\alias{viewOverlay}
\title{viewOverlay}
\usage{
viewOverlay(
  data,
  proc.img.dir,
  well.label,
  obj.label,
  obj.color = NULL,
  obj.col.pal = NULL,
  obj.shape = NULL,
  obj.shape.pal = NULL,
  text.anno = NULL,
  file = NULL
)
}
\arguments{
\item{data}{A data frame output from \code{modelSelection} or any \code{OF} function from easyXpress.
The dataframe should be filtered to contain only the wells to be plotted. Please Arrange the dataframe in the order that overlays are to be plotted.}

\item{proc.img.dir}{A variable name in \code{data} that holds the full PATH to the directory holding processed images matching the data.
The processed images must have the standard \code{_overlay.png} suffix and file name output from CellProfiler.}

\item{well.label}{A variable name in \code{data} to display as a well label. For example, \code{"Metadata_Well"}.}

\item{obj.label}{A variable name in \code{data} to label objects by. For example, \code{"model"}.}

\item{obj.color}{Optional: A variable name in \code{data} to color objects by. The default is NULL.}

\item{obj.col.pal}{Optional: A color palette for obj.labels. This is a vector of colors with names for each unique value in \code{obj.color}. NULL passes color values to \code{obj.label}.}

\item{obj.shape}{Optional: A variable name in \code{data} to use for the shape of objects. The default is NULL.}

\item{obj.shape.pal}{Optional: A shape palette for objects. This is a vector of shape values with names for each unique value in \code{obj.shape}. NULL passes shape values to \code{obj.label}.}

\item{text.anno}{Optional: a variable name in \code{data} used to display text annotation at an object center.
by default this value is shifted vertically to avoid plotting over object center.}

\item{file}{Optional: The full path for saving output plot. Default is NULL which will just return the ggplot2 plot.}
}
\value{
A plot showing all the CellProfiler processed well overlays in \code{data} with objects annotated as desired.
}
\description{
viewOverlay
}
