% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewWell.R
\name{viewWell}
\alias{viewWell}
\title{viewWell}
\usage{
viewWell(df, img_dir, plate, well, boxplot = TRUE)
}
\arguments{
\item{df}{Unsummarised CellProfiler data output from the
\code{Xpress} functon. Either the raw or processed data can be viewed.}

\item{img_dir}{The full path of the directory holding
the processed images.}

\item{plate}{Desired plate to be analyzed.}

\item{well}{Desired well to be analyzed.}

\item{boxplot}{should output include boxplot of object data?}
}
\value{
A ggplot object with selected well image and boxplot.
}
\description{
\code{viewWell} plots the processed well image
with object centroids colored by type.
}
