% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierOF.R
\name{classifierOF}
\alias{classifierOF}
\title{classifierOF}
\usage{
classifierOF(data, model = "gbm2x")
}
\arguments{
\item{data}{A data frame output from the \code{modelSelection} function.}

\item{model}{Specify one of the built in models. Currently only "gbm2x" is available. This classifier was
trained on over 1,000 worm objects selected from a large GWAS experiment and classifies objects as "worm" or "non-worm" with ~90% accuracy.
NOTE: The model was trained to classify poorly segmented worms as 'non-worm' so true worms are often classified as "non-worm".}
}
\value{
A single data frame identical to the input data with the \code{classifier_ObjectFlag} variable added.
The \code{classifier_ObjectFlag} variable is coded as \code{"classifier"} for objects that are called non-worm by the 2X classifier.
All other objects are coded as \code{NA_character}. The \code{gbm2x_worm_prob} variable provides the probability that the object is a properly segmented worm.
}
\description{
This function will flag non-worm objects using machine learning.
}
