% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeFlag.R
\name{edgeFlag}
\alias{edgeFlag}
\title{edgeFlag}
\usage{
edgeFlag(model_selected_data, radius = 825)
}
\arguments{
\item{model_selected_data}{A data frame where selected worm model
denotations have been joined and assigned to each worm object}

\item{radius}{Radius in pixels away from image center with even illumination.}
}
\value{
A single data frame with worm objects on the edge
of the well identified but retained
}
\description{
This is the primary function for flagging well edge worms
in CellProfiler data
}
