% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXpress.R
\name{readXpress}
\alias{readXpress}
\title{readXpress}
\usage{
readXpress(filedir, design = FALSE, rdafile = NULL)
}
\arguments{
\item{filedir}{The directory with CellProfiler data.
This directory should have CellProfiler data in a sub-folder named cp_data.
The data can be in .Rda, .Rds, or .csv formats.}

\item{design}{Logical parameter, if TRUE then a design file
will be joined to data.
The design file should be located in a sub-folder
of the filedir named design.
If FALSE no design file will be joined.}

\item{rdafile}{The file name of a particular .rda file
to load from the cp_data sub-folder.}
}
\value{
A single data frame named raw_data that contains
all CellProfiler model outputs as well as experimental treatments
if a design file is used.
}
\description{
This is the primary fuction for reading CellProfiler data
into R with this package.
It is built exclusively for use with worm image data.
}
