% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyImages.R
\name{tidyImages}
\alias{tidyImages}
\title{tidyImages}
\usage{
tidyImages(project_dir, rm_other = FALSE)
}
\arguments{
\item{project_dir}{the full path to the project directory. This folder should
contain .TIF images exported from the imaging platform.}

\item{rm_other}{logical, remove non-TIF files and folders after the .TIF
files are moved to raw_images subfolder? Default is \code{FALSE}. Set to
\code{TRUE} if you are certain you want to remove non-TIF files and folders
in the \code{project_dir} after .TIFs are moved.}
}
\value{
A folder named raw_images under the directory specified with
  \code{project_dir}, which contains all .TIF image files for the project.
}
\description{
Organize .TIF files exported from an imager platform by creating a raw_images
subdirectory in the project directory, moving all .TIF files into it, and
removing non-TIF files and folders if desired.
}
