% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewPlate.R
\name{viewPlate}
\alias{viewPlate}
\title{viewPlate}
\usage{
viewPlate(df, plate)
}
\arguments{
\item{df}{Summarised CellProfiler data output from the
\code{Xpress} or \code{process} functons. Either the raw or processed data can be viewed.}

\item{plate}{Desired plate to be analyzed.}
}
\value{
A ggplot object with selected plate.
}
\description{
\code{viewPlate} plots the processed plate.
}
