% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xpress.R
\name{Xpress}
\alias{Xpress}
\title{Xpress}
\usage{
Xpress(
  filedir,
  rdafile,
  ...,
  design = FALSE,
  radius = 825,
  center_x = 1024,
  center_y = 1024,
  cluster_flag = TRUE,
  well_edge_flag = TRUE
)
}
\arguments{
\item{filedir}{The project directory or directories with CellProfiler data.
Provide a full path to the directory or a vector of project directory paths.
Each directory must have a \code{cellprofiler-nf} output .rda file in a sub-folder named \code{cp_data}.}

\item{rdafile}{The specific .rda file name in the \code{cp_data} directory to read.
If multiple project directories are supplied to \code{filedir},
then include the .rda files for each project in the same order of the directories given in \code{filedir}.}

\item{...}{specify variable used to summarize data. Variable names can be listed in succession.}

\item{design}{Logical parameter, if TRUE then a design file
will be joined to data.
The design file should be located in a sub-folder
of the filedir named design.
If FALSE no design file will be joined.}

\item{radius}{Radius in pixels away from image center with even illumination.}

\item{center_x}{center x position of image}

\item{center_y}{center y position of image}

\item{cluster_flag}{Logical parameter; do you want worm objects
in a cluster to be excluded when calculating well outliers?
We recommend TRUE as the default.}

\item{well_edge_flag}{Logical parameter; do you want worm objects
in close proximity to the well edge to be excluded
when calculating well outliers? We recommend TRUE as the default.}
}
\value{
List with four elements: raw data, processed data, and summaries for both datasets.
}
\description{
This is a wrapper function that will run all functions in the package
and output a processed dataframe.
}
