% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xpress.R
\name{Xpress}
\alias{Xpress}
\title{Xpress}
\usage{
Xpress(
  filedir,
  design = FALSE,
  radius = 825,
  cluster_flag = TRUE,
  well_edge_flag = TRUE,
  ...
)
}
\arguments{
\item{filedir}{The directory with CellProfiler data.
This directory should have CellProfiler data in a sub-folder named cp_data.
The data can be in .Rda, .Rds, or .csv formats.}

\item{design}{Logical parameter, if TRUE then a design file
will be joined to data.
The design file should be located in a sub-folder
of the filedir named design.
If FALSE no design file will be joined.}

\item{radius}{Radius in pixels away from image center with even illumination.}

\item{cluster_flag}{Logical parameter; do you want worm objects
in a cluster to be excluded when calculating well outliers?
We recommend TRUE as the default.}

\item{well_edge_flag}{Logical parameter; do you want worm objects
in close proximity to the well edge to be excluded
when calculating well outliers? We recommend TRUE as the default.}

\item{...}{specify variable used to summarize data}
}
\value{
List with four elements: raw data, processed data, and summaries for both datasets.
}
\description{
This is a wrapper function that will run all functions in the package
and output a processed dataframe.
}
